requirements.txt
flask
flask-login
flask-wtf
wtforms
werkzeug
pandas
tabula-py
reportlab
gunicorn
email-validator
